A3/requirements.txt
# A3 pipeline — pinned for reproducible Docker builds
asyncpg==0.29.0
matplotlib==3.8.3
opentelemetry-api==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-exporter-otlp-proto-grpc==1.23.0
prometheus-client==0.20.0
psutil==5.9.8
psycopg[binary]==3.1.18
